+++ Tests/Validator/Constraints/MoneyRangeValidatorTest.php
@@ -2,22 +2,22 @@
 
 namespace NetTeam\Bundle\DDDBundle\Tests\Validator\Constraints;
 
-use NetTeam\Bundle\DDDBundle\Validator\Constraints\MoneyRange as MoneyRangeConstraint;
-use NetTeam\Bundle\DDDBundle\Validator\Constraints\MoneyRangeValidator;
+use NetTeam\Bundle\DDDBundle\Validator\Constraints\MoneyRangeValue as MoneyRangeConstraint;
+use NetTeam\Bundle\DDDBundle\Validator\Constraints\MoneyRangeValueValidator;
 use NetTeam\DDD\ValueObject\MoneyRange;
 use NetTeam\DDD\ValueObject\Money;
 
 /**
  * @group Unit
  */
-class MoneyRangeValidatorTest extends \PHPUnit_Framework_TestCase
+class MoneyRangeValueValidatorTest extends \PHPUnit_Framework_TestCase
 {
     private $validator;
     private $context;
 
     protected function setUp()
     {
-        $this->validator = new MoneyRangeValidator();
+        $this->validator = new MoneyRangeValueValidator();
         $this->context = $this->getMock('Symfony\Component\Validator\ExecutionContext', array(), array(), '', false);
         $this->validator->initialize($this->context);
     }